backendmotel/requirement.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-appconf==1.0.4
django-filter==2.3.0
django-imagekit==4.0.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
googlemaps==4.4.1
idna==2.10
Markdown==3.2.2
phonenumbers==8.12.6
pilkit==2.0
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
undefined==0.0.8
